License the repo under the BSD-3-Clause license (#186)
I also took the occasion to change license headers to SPDX-style tags (see https://spdx.github.io/spdx-spec/v2.3/file-tags/), that are more manageable and machine readable. 

As a result of the cleanup I deleted the authors name from the headers as it is more mantainable to just store this information in git, if there is any problem in this we can re-introduce them with a follow-up PR, using the `SPDX-FileContributor:` tag.